requirements.txt
# Core Dependencies
gradio>=4.0.0,<6.0.0
pandas>=1.5.0
numpy>=1.21.0
plotly>=5.0.0

# Analysis Libraries
causalnex>=0.12.0
scikit-learn>=1.0.0
statsmodels>=0.13.0
networkx>=2.8.0
scipy>=1.9.0

# Data Processing
openpyxl>=3.0.0  # Excel file support
xlrd>=2.0.0      # Excel file support
pyarrow>=10.0.0  # Parquet file support
requests>=2.28.0 # URL loading support

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# Utilities
python-dateutil>=2.8.0
pytz>=2022.1

# Development Dependencies (optional)
pytest>=7.0.0
black>=22.0.0
flake8>=5.0.0
mypy>=0.991
pywebview>=4.0.0
